bip32: Add serde for Extended(Priv|Pub)Key
